+++ src/PeNet/Structures/IMAGE_NT_HEADERS.cs
@@ -24,7 +24,7 @@ namespace PeNet.Structures
         /// </summary>
         /// <param name="peFile">A PE file .</param>
         /// <param name="offset">Raw offset of the NT header.</param>
-        public IMAGE_NT_HEADERS(IRawFile peFile, int offset)
+        public IMAGE_NT_HEADERS(IRawFile peFile, long offset)
             : base(peFile, offset)
         {
             FileHeader = new IMAGE_FILE_HEADER(peFile, offset + 0x4);
